requirements.txt
websockets==12.0
fastapi==0.109.2
uvicorn==0.27.1
pyserial
